chore: reimplement blacklists
`dmpconfig.json` was being imported using `require()` this means that
the content the file must be known at (what basically is) compile-time
which is hard to do when we want to also manage a list of strings in it.

If we want minimal set-up then `dmpconfig.json` needs to already exist
when you've cloned the repo, the problem is that if we do that that
means that we have a file that **will** be overwritten upon updating,
which means that we can't store dynamic information like the blacklist
in it.

To resolve this, we remove the constants that would have otherwise been
in `dmpconfg.json` to `config.json` and then dynamically write
`blacklist.json` in `backups/` instead.

it's 3:00am, I have work tomorrow lol